# Faster metadata hashing