@esben 2016-03-29
Last patch is fine.  The entire code is not really so pretty, so it just adapts to the style ;)